grad1/requirements.txt
alembic==1.1.0
altgraph==0.16.1
aniso8601==8.0.0
Appium-Python-Client==0.48
asn1crypto==0.24.0
attrs==19.3.0
Automat==0.8.0
Babel==2.7.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.8.0
bleach==3.1.0
blinker==1.4
brotlipy==0.7.0
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.1
comtypes==1.1.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.3.1
cssselect==1.1.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.5.0
diff-match-patch==20121119
django==2.2.1
django-cors-headers==3.1.1
django-crispy-forms==1.7.2
django-filter==2.2.0
django-formtools==2.1
django-guardian==1.4.9
django-haystack==2.8.1
django-import-export==1.0.0
django-reversion==2.0.13
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-jwt-proxy==0.1
drf-extensions==0.3.1
entrypoints==0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Faker==2.0.3
Flask==1.1.1
Flask-Admin==1.5.3
Flask-BabelEx==0.9.3
Flask-Cache==0.13.1
Flask-Cors==3.0.8
Flask-HTTPAuth==3.3.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
fonttools==4.0.2
future==0.16.0
geetest==3.2.1
h11==0.7.0
h2==3.1.1
hpack==3.0.0
html5lib==1.0.1
httplib2==0.9.2
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.6
imageio==2.5.0
imageio-ffmpeg==0.3.0
importlib-metadata==0.23
incremental==17.5.0
ipykernel==5.1.3
ipython==7.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itchat==1.2.32
itsdangerous==1.1.0
itypes==1.1.0
jdcal==1.3
jedi==0.15.1
jieba==0.39
Jinja2==2.10
jsonschema==3.1.1
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.0.0
jupyter-core==4.6.1
jwt==0.5.3
kaitaistruct==0.8
kiwisolver==1.1.0
ldap3==2.5.2
lml==0.0.9
lxml==4.4.1
Mako==1.1.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==3.1.1
mistune==0.8.4
mitmproxy==4.0.4
more-itertools==7.2.0
moviepy==1.0.1
MyQR==2.3.1
nbconvert==5.6.1
nbformat==4.4.0
notebook==6.0.2
numpy==1.17.0
oauthlib==2.0.7
odfpy==1.3.6
opencv-python==4.1.1.26
openpyxl==2.5.2
pandas==0.25.3
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.1
passlib==1.7.1
pefile==2019.4.18
pickleshare==0.7.5
Pillow==5.1.0
prettytable==0.7.2
proglog==0.1.9
prometheus-client==0.7.1
prompt-toolkit==2.0.10
Protego==0.1.15
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycparser==2.18
PyDispatcher==2.0.5
pydivert==2.1.0
pyecharts==1.4.0
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
pygame==1.9.6
Pygments==2.4.2
PyHamcrest==1.9.0
PyInstaller==3.5
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL==18.0.0
pyparsing==2.2.2
pyperclip==1.6.5
pypng==0.0.20
PyQRCode==1.2.1
pyquery==1.4.0
pyrsistent==0.15.5
pytesseract==0.3.0
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.2
pywifi==1.1.8
pywin32==225
pywin32-ctypes==0.2.0
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.1.0
qtconsole==4.5.5
queuelib==1.5.0
redis==3.3.8
requests==2.22.0
rsa==4.0
ruamel.yaml==0.15.100
Scrapy==1.8.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
setup.py==0.1
simplejson==3.16.0
six==1.10.0
sortedcontainers==2.0.5
soupsieve==1.9.4
speaklater==1.3
SQLAlchemy==1.3.8
sqlparse==0.3.0
tablib==0.13.0
terminado==0.8.2
testpath==0.4.4
text-unidecode==1.3
tornado==5.1.1
tqdm==4.42.1
traitlets==4.3.3
Twisted==19.7.0
uritemplate==3.0.0
urllib3==1.25.6
urwid==2.0.1
virtualenv==16.7.4
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.6
Whoosh==2.7.4
widgetsnbextension==3.5.1
wsproto==0.11.0
WTForms==2.2.1
wxpy==0.3.9.8
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
zope.interface==4.6.0
